Adopt LuaJIT rolling release changes
LuaJIT recently switched to the rolling release scheme:

* 2.0.x -> 2.0.ROLLING
* 2.1.x -> 2.1.ROLLING

See LuaJIT commits:

* 50e0fa03c48cb9af03c3efdc3100f12687651a2e
* 2090842410e0ba6f81fad310a77bf5432488249a

In addition to the changing of the versioning scheme, there has been
introduced some changes in the MinGW build.

See the commit and the PR:

* 19707009bfb8d1fe59a5c328034e8e8ad1b56232
* https://github.com/LuaJIT/LuaJIT/pull/1071

The problem is that if sh is in PATH, mingw32-make will use sh instead
of cmd.exe, but some recipes in Makefile use cmd.exe syntax if MinGW is
detected.

With this patch, we force mingw32-make to use Windows shell (SHELL=cmd).

Fixes: https://github.com/luarocks/hererocks/issues/25